backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
redis
pydantic
pydantic-settings
python-jose[cryptography]
cryptography
firebase-admin
httpx
numpy
pandas
websockets
apscheduler
python-dotenv
MetaTrader5
